Replaced uint32_t with size_t.

+++ src/drivers/fdc.c
@@ -82,7 +82,7 @@ int fdc_is_busy()
     return 0;
 }
 
-uint32_t fdc_read_status()
+size_t fdc_read_status()
 {
     return inb(FDC_MSR);
 }
@@ -110,7 +110,7 @@ void fdc_write_ccr(uint8_t val)
 
 void fdc_read_sector_imp (uint8_t head, uint8_t track, uint8_t sector) {
 
-    uint32_t st0, cyl;
+    size_t st0, cyl;
 
     //! set the DMA for read transfer
     fdc_init_dma_read ();
@@ -141,9 +141,9 @@ void fdc_read_sector_imp (uint8_t head, uint8_t track, uint8_t sector) {
     fdc_check_int (&st0,&cyl);
 }
 
-void fdc_drive_data (uint32_t stepr, uint32_t loadt, uint32_t unloadt, bool dma ) {
+void fdc_drive_data (size_t stepr, size_t loadt, size_t unloadt, bool dma ) {
 
-    uint32_t data = 0;
+    size_t data = 0;
 
     fdc_send_command (FDC_CMD_SPECIFY);
 
@@ -154,9 +154,9 @@ void fdc_drive_data (uint32_t stepr, uint32_t loadt, uint32_t unloadt, bool dma 
     fdc_send_command (data);
 }
 
-int fdc_calibrate (uint32_t drive) {
+int fdc_calibrate (size_t drive) {
 
-    uint32_t st0, cyl;
+    size_t st0, cyl;
 
     if (drive >= 4)
         return -2;
@@ -184,7 +184,7 @@ int fdc_calibrate (uint32_t drive) {
     return -1;
 }
 
-void fdc_check_int (uint32_t* st0, uint32_t* cyl) {
+void fdc_check_int (size_t* st0, size_t* cyl) {
 
     fdc_send_command (FDC_CMD_CHECK_INT);
 
@@ -192,9 +192,9 @@ void fdc_check_int (uint32_t* st0, uint32_t* cyl) {
     *cyl = fdc_read_data ();
 }
 
-int fdc_seek ( uint32_t cyl, uint32_t head ) {
+int fdc_seek ( size_t cyl, size_t head ) {
 
-    uint32_t st0, cyl0;
+    size_t st0, cyl0;
 
     if (fdd >= 4)
         return -1;
@@ -230,7 +230,7 @@ void fdc_enable_controller () {
 
 void fdc_reset () {
 
-    uint32_t st0, cyl;
+    size_t st0, cyl;
 
     //! reset the controller
     fdc_disable_controller ();
