+++ src/fs/initrd.c
@@ -19,7 +19,7 @@ inode_t *initrd_nodes;
 extern mnt_table_t mounted_fs[256];
 
 /* Initrd File System */
-void initrd_load(uint32_t location)
+void initrd_load(size_t location)
 {
     initrd_sb = (initrd_super_block_t*) location;
     fheader = NULL;
@@ -30,7 +30,7 @@ void initrd_load(uint32_t location)
     for (int i = 0; i < initrd_sb->files; i++) {
         fheader = (initrd_file_header_t*) location;
         memcpy(&(initrd_headers[i]), fheader, sizeof(initrd_file_header_t));
-        uint32_t block_addr = (uint32_t) fheader + initrd_sb->fh_struct_size;
+        size_t block_addr = (size_t) fheader + initrd_sb->fh_struct_size;
         inode_init(&initrd_nodes[i], fheader->flags, fheader->inode,
                     fheader->length, 0, 0, block_addr, INITRD_TYPE);
         initrd_nodes[i].open = &initrd_open;
@@ -38,8 +38,8 @@ void initrd_load(uint32_t location)
         initrd_nodes[i].rewind = &initrd_rewind;
         initrd_nodes[i].close = &initrd_close;
 
-        location += (uint32_t) sizeof(initrd_file_header_t);
-        location += (uint32_t) fheader->length;
+        location += (size_t) sizeof(initrd_file_header_t);
+        location += (size_t) fheader->length;
     }
 }
 
@@ -69,8 +69,8 @@ void initrd_mount(super_block_t *sb, int index, char *path)
 
     for (int i = 1; i < sb_data->files; i++) {
         struct dentry *temp_de = get_dentry_by_inode(&initrd_nodes[i]);
-        uint32_t addr = (uint32_t) temp_de;
-        inode_write(node, sizeof(uint32_t), 1, (char*) &addr);
+        size_t addr = (size_t) temp_de;
+        inode_write(node, sizeof(size_t), 1, (char*) &addr);
 
         graph_node_t *temp_node = graph_create((void*) temp_de);
         graph_add_node(path_node, temp_node);
@@ -79,9 +79,9 @@ void initrd_mount(super_block_t *sb, int index, char *path)
 
 int initrd_read(struct inode *node, size_t sz, int n, char *buf)
 {
-    uint32_t i;
-    uint32_t phys = node->block + node->offset;
-    uint32_t limit = node->block + node->length;
+    size_t i;
+    size_t phys = node->block + node->offset;
+    size_t limit = node->block + node->length;
 
     for (i = 0; i < sz * n && phys + i < limit; i++) {
         buf[i] = ((char*) phys)[i];
@@ -90,7 +90,7 @@ int initrd_read(struct inode *node, size_t sz, int n, char *buf)
     return i;
 }
 
-int initrd_open(struct inode *node, uint32_t flags)
+int initrd_open(struct inode *node, size_t flags)
 {
     return -1;
 }
