+++ src/fs/procfs.c
@@ -22,7 +22,7 @@ void procfs_init()
     procfs_sb.fh_struct_size = sizeof(procfs_file_header_t);
 }
 
-struct inode *procfs_alloc_inode(uint32_t flags, uint32_t length)
+struct inode *procfs_alloc_inode(size_t flags, size_t length)
 {
     struct inode *node = kmalloc(sizeof(struct inode), 0, 0);
     inode_init(node, flags, inode_id++, length, 0, 0, 0, PROCFS_TYPE);
@@ -37,7 +37,7 @@ struct inode *procfs_alloc_inode(uint32_t flags, uint32_t length)
     return node;
 }
 
-void procfs_alloc(struct inode *node, uint32_t length)
+void procfs_alloc(struct inode *node, size_t length)
 {
     if (!node->block || !node->length) {
         node->block = kmalloc(length + procfs_sb.fh_struct_size, 0, 0);
@@ -54,7 +54,7 @@ void procfs_alloc(struct inode *node, uint32_t length)
     addr->length = length;
 }
 
-int procfs_create(struct inode *parent, char *name, uint32_t flags)
+int procfs_create(struct inode *parent, char *name, size_t flags)
 {
     struct dentry *parent_de = get_dentry_by_inode(parent);
     if (parent_de == NULL) {
@@ -73,16 +73,16 @@ int procfs_create(struct inode *parent, char *name, uint32_t flags)
     new_de->inode = new_inode;
     add_dentry(new_de);
 
-    uint32_t addr = &new_de;
+    size_t addr = &new_de;
     inode_rewind(parent);
-    uint32_t temp_addr;
-    inode_read(parent, sizeof(uint32_t), 1, (char*) temp_addr);
+    size_t temp_addr;
+    inode_read(parent, sizeof(size_t), 1, (char*) temp_addr);
     while (temp_addr != 0x0) {
-        inode_read(parent, sizeof(uint32_t), 1, (char*) temp_addr);
+        inode_read(parent, sizeof(size_t), 1, (char*) temp_addr);
     }
-    parent->offset += sizeof(uint32_t);
+    parent->offset += sizeof(size_t);
 
-    inode_write(parent, sizeof(uint32_t), 1, (char*) addr);
+    inode_write(parent, sizeof(size_t), 1, (char*) addr);
 
     graph_node_t *new_node = graph_create((void*) new_de);
     graph_node_t *parent_node = get_node_by_path(cwd);
@@ -93,9 +93,9 @@ int procfs_create(struct inode *parent, char *name, uint32_t flags)
 
 int procfs_write(struct inode *node, size_t sz, int n, char *buf)
 {
-    uint32_t i;
-    uint32_t phys = node->block + node->offset + procfs_sb.fh_struct_size;
-    uint32_t limit = node->block + node->length + procfs_sb.fh_struct_size;
+    size_t i;
+    size_t phys = node->block + node->offset + procfs_sb.fh_struct_size;
+    size_t limit = node->block + node->length + procfs_sb.fh_struct_size;
 
     if (limit - phys < sz * n) {
         procfs_alloc(node, node->length + sz * n);
@@ -112,9 +112,9 @@ int procfs_write(struct inode *node, size_t sz, int n, char *buf)
 
 int procfs_read(struct inode *node, size_t sz, int n, char *buf)
 {
-    uint32_t i;
-    uint32_t phys = node->block + node->offset + procfs_sb.fh_struct_size;
-    uint32_t limit = node->block + node->length + procfs_sb.fh_struct_size;
+    size_t i;
+    size_t phys = node->block + node->offset + procfs_sb.fh_struct_size;
+    size_t limit = node->block + node->length + procfs_sb.fh_struct_size;
 
     for (i = 0; i < sz * n && phys + i < limit; i++) {
         buf[i] = ((char*) phys)[i];
@@ -123,7 +123,7 @@ int procfs_read(struct inode *node, size_t sz, int n, char *buf)
     return i;
 }
 
-int procfs_open(struct inode *node, uint32_t flags)
+int procfs_open(struct inode *node, size_t flags)
 {
     return 0;
 }
