+++ src/include/fs/vfs.h
@@ -37,38 +37,38 @@ struct stat;
 typedef int (*inode_read_t)(struct inode *node, size_t sz, int n, char *buf);
 typedef int (*inode_write_t)(struct inode *node, size_t sz, int n, char *buf);
 typedef int (*inode_rewind_t)(struct inode *node);
-typedef int (*inode_open_t)(struct inode *node, uint32_t flags);
+typedef int (*inode_open_t)(struct inode *node, size_t flags);
 typedef int (*inode_close_t)(struct inode *node);
 typedef struct dentry *(*inode_readdir_t)(struct inode *node);
 typedef struct inode *(*inode_finddir_t)(struct inode *node, char *name);
-typedef int (*inode_create_t)(struct inode *parent, char *name, uint32_t flags);
-typedef int (*inode_mkdir_t)(struct inode *parent, char *name, uint32_t flags);
+typedef int (*inode_create_t)(struct inode *parent, char *name, size_t flags);
+typedef int (*inode_mkdir_t)(struct inode *parent, char *name, size_t flags);
 typedef int (*inode_ioctl_t)(struct inode *node, int req, void *argp);
 typedef size_t (*inode_getsize_t)(struct inode *node);
-typedef int (*inode_chmod_t)(struct inode *node, uint32_t mode);
+typedef int (*inode_chmod_t)(struct inode *node, size_t mode);
 typedef int (*inode_unlink_t)(struct inode *node);
 typedef int (*inode_symlink_t)(struct inode *node, struct inode *link);
 typedef int (*inode_readlink_t)(struct inode *node);
 
 typedef struct inode {
     void *device;         // Device where node is found
-    uint32_t fsid;        // Filesystem type id
-    uint32_t mask;        // Node mask
-    uint32_t uid;         // User ID
-    uint32_t gid;         // Group ID
-    uint32_t flags;       // Node flags
-    uint32_t inode;       // Node ID
+    size_t fsid;        // Filesystem type id
+    size_t mask;        // Node mask
+    size_t uid;         // User ID
+    size_t gid;         // Group ID
+    size_t flags;       // Node flags
+    size_t inode;       // Node ID
     size_t length;        // Node content length
-    uint32_t nblocks;     // Number of blocks
-    uint32_t blocksz;     // Block size
-    uint32_t block;       // Phys. block
-    uint32_t open_flags;  // Flags for open
-    uint32_t lock;        // Inode is locked or not
+    size_t nblocks;     // Number of blocks
+    size_t blocksz;     // Block size
+    size_t block;       // Phys. block
+    size_t open_flags;  // Flags for open
+    size_t lock;        // Inode is locked or not
     size_t offset;        // Cursor position
 
-    uint32_t atime;       // Accessed time
-    uint32_t mtime;       // Modified time
-    uint32_t ctime;       // Created time
+    size_t atime;       // Accessed time
+    size_t mtime;       // Modified time
+    size_t ctime;       // Created time
 
     // Functions
     inode_read_t      read;
@@ -90,7 +90,7 @@ typedef struct inode {
     // Link and reference
     struct inode *link;
     int32_t refcount;
-    uint32_t nlink;
+    size_t nlink;
 } inode_t;
 
 typedef struct {
@@ -100,7 +100,7 @@ typedef struct {
 typedef struct super_block {
     void *device;
     inode_t *inode;
-    uint32_t block;
+    size_t block;
     file_system_type *fs_type;
     void *fs_data;
 
@@ -162,18 +162,18 @@ struct file_operations {
 struct stat {
     uint16_t st_dev;
     uint16_t st_ino;
-    uint32_t st_mode;
+    size_t st_mode;
     uint16_t st_nlink;
     uint16_t st_uid;
     uint16_t st_gid;
     uint16_t st_rdev;
-    uint32_t st_size;
-    uint32_t st_atime;
-    uint32_t __unused1;
-    uint32_t st_mtime;
-    uint32_t __unused2;
-    uint32_t st_ctime;
-    uint32_t __unused3;
+    size_t st_size;
+    size_t st_atime;
+    size_t __unused1;
+    size_t st_mtime;
+    size_t __unused2;
+    size_t st_ctime;
+    size_t __unused3;
 };
 
 typedef struct path_tokens {
@@ -192,20 +192,20 @@ struct dentry *get_dentry_by_tokens(path_tokens *tokens);
 path_tokens *tokenize_path(char *path);
 void destroy_tokens(path_tokens *tokens);
 
-int inode_init(struct inode *node, uint32_t flags, uint32_t id, uint32_t length,
-                uint32_t nblocks, uint32_t blocksz, uint32_t block, uint32_t fsid);
+int inode_init(struct inode *node, size_t flags, size_t id, size_t length,
+                size_t nblocks, size_t blocksz, size_t block, size_t fsid);
 int inode_read(struct inode *node, size_t sz, int n, char *buf);
 int inode_write(struct inode *node, size_t sz, int n, char *buf);
 int inode_rewind(struct inode *node);
-int inode_open(struct inode *node, uint32_t flags);
+int inode_open(struct inode *node, size_t flags);
 int inode_close(struct inode *node);
 struct dentry *inode_readdir(struct inode *node);
 struct inode *inode_finddir(struct inode *node, char *name);
-int inode_create(struct inode *node, char *name, uint32_t flags);
-int inode_mkdir(struct inode *node, char *name, uint32_t flags);
+int inode_create(struct inode *node, char *name, size_t flags);
+int inode_mkdir(struct inode *node, char *name, size_t flags);
 int inode_ioctl(struct inode *node, int req, void *argp);
 size_t inode_getsize(struct inode *node);
-int inode_chmod(struct inode *node, uint32_t mode);
+int inode_chmod(struct inode *node, size_t mode);
 int inode_unlink(struct inode *node);
 int inode_symlink(struct inode *node, struct inode *link);
 int inode_readlink(struct inode *node);
